+++ migrations/001_assessement_access_rules__time_limit_min__add.sql
@@ -1,0 +1,1 @@
+ALTER TABLE assessment_access_rules ADD COLUMN IF NOT EXISTS time_limit_min integer;