+++ models/assessment_questions.sql
@@ -15,7 +15,3 @@ CREATE TABLE IF NOT EXISTS assessment_questions (
 
 CREATE INDEX IF NOT EXISTS assessment_questions_assessment_id_idx ON assessment_questions (assessment_id);
 CREATE INDEX IF NOT EXISTS assessment_questions_alternative_group_id_idx ON assessment_questions (alternative_group_id);
-
-
-ALTER TABLE assessment_questions ALTER COLUMN alternative_group_id DROP NOT NULL;
-ALTER TABLE assessment_questions ADD COLUMN IF NOT EXISTS force_max_points BOOLEAN DEFAULT FALSE;