+++ models/enum_role_add_none.sql
@@ -1,1 +1,0 @@
-ALTER TYPE enum_role ADD VALUE IF NOT EXISTS 'None' BEFORE 'Student';