+++ models/instance_questions.sql
@@ -21,22 +21,3 @@ CREATE TABLE IF NOT EXISTS instance_questions (
 );
 
 CREATE INDEX IF NOT EXISTS instance_questions_assessment_instance_id_idx ON instance_questions (assessment_instance_id);
-
-ALTER TABLE instance_questions ADD COLUMN IF NOT EXISTS status enum_instance_question_status DEFAULT 'unanswered';
-
-ALTER TABLE instance_questions ADD COLUMN IF NOT EXISTS authn_user_id bigint;
-
-DO $$
-BEGIN
-    IF NOT EXISTS (
-            SELECT 1 FROM information_schema.table_constraints
-            WHERE constraint_name = 'instance_questions_authn_user_id_fkey'
-        )
-        THEN
-        ALTER TABLE instance_questions ADD FOREIGN KEY (authn_user_id) REFERENCES users ON DELETE CASCADE ON UPDATE CASCADE;
-    END IF;
-END;
-$$;
-
-ALTER TABLE instance_questions ADD COLUMN IF NOT EXISTS duration INTERVAL DEFAULT INTERVAL '0 seconds';
-ALTER TABLE instance_questions ADD COLUMN IF NOT EXISTS first_duration INTERVAL;