+++ models/job_sequences.sql
@@ -16,31 +16,3 @@ CREATE TABLE IF NOT EXISTS job_sequences (
 
 CREATE INDEX IF NOT EXISTS job_sequences_assessment_id_idx ON job_sequences (assessment_id);
 CREATE INDEX IF NOT EXISTS job_sequences_course_id_idx ON job_sequences (course_id);
-
-ALTER TABLE job_sequences ADD COLUMN IF NOT EXISTS course_instance_id BIGINT REFERENCES course_instances ON DELETE CASCADE ON UPDATE CASCADE;
-
-DO $$
-BEGIN
-    IF NOT EXISTS (
-            SELECT 1 FROM information_schema.table_constraints
-            WHERE constraint_name = 'job_sequences_course_instance_id_fkey'
-        )
-        THEN
-        ALTER TABLE job_sequences ADD FOREIGN KEY (course_instance_id) REFERENCES course_instances ON DELETE CASCADE ON UPDATE CASCADE;
-    END IF;
-END;
-$$;
-
-ALTER TABLE job_sequences ADD COLUMN IF NOT EXISTS assessment_id BIGINT REFERENCES assessments ON DELETE CASCADE ON UPDATE CASCADE;
-
-DO $$
-BEGIN
-    IF NOT EXISTS (
-            SELECT 1 FROM information_schema.table_constraints
-            WHERE constraint_name = 'job_sequences_assessment_id_fkey'
-        )
-        THEN
-        ALTER TABLE job_sequences ADD FOREIGN KEY (assessment_id) REFERENCES assessments ON DELETE CASCADE ON UPDATE CASCADE;
-    END IF;
-END;
-$$;