+++ models/pl_courses.sql
@@ -9,6 +9,4 @@ CREATE TABLE IF NOT EXISTS pl_courses (
     deleted_at TIMESTAMP WITH TIME ZONE
 );
 
-ALTER TABLE pl_courses ADD COLUMN IF NOT EXISTS display_timezone text;
-UPDATE pl_courses SET display_timezone = 'America/Chicago' WHERE display_timezone IS NULL;
 -- FIXME: make display_timezone NOT NULL in the future