+++ models/questions.sql
@@ -20,21 +20,3 @@ CREATE TABLE IF NOT EXISTS questions (
 );
 
 CREATE INDEX IF NOT EXISTS questions_topic_id_idx ON questions (topic_id);
-
-ALTER TABLE questions DROP CONSTRAINT IF EXISTS questions_number_course_id_key;
-
-DO $$
-BEGIN
-    IF NOT EXISTS (
-            SELECT 1 FROM information_schema.table_constraints
-            WHERE constraint_name = 'questions_course_id_number_key'
-        )
-        THEN
-        ALTER TABLE questions ADD UNIQUE (course_id, number);
-    END IF;
-END;
-$$;
-
-ALTER TABLE questions ADD COLUMN IF NOT EXISTS autograding_enabled BOOLEAN DEFAULT FALSE;
-ALTER TABLE questions ADD COLUMN IF NOT EXISTS autograder text;
-ALTER TABLE questions ADD COLUMN IF NOT EXISTS environment text;