+++ models/submissions.sql
@@ -19,6 +19,3 @@ CREATE TABLE IF NOT EXISTS submissions (
 );
 
 CREATE INDEX IF NOT EXISTS submissions_variant_id_idx ON submissions (variant_id);
-ALTER TABLE submissions ADD COLUMN IF NOT EXISTS grading_method enum_grading_method;
-
-ALTER TABLE submissions ADD COLUMN IF NOT EXISTS duration INTERVAL DEFAULT INTERVAL '0 seconds';